31/05/2026

+++ GymMachineWS/DataAccessLayer/Creator/OrderCreator.cs
@@ -10,8 +10,8 @@ namespace GymMachineWS.DataAccessLayer
         {
             return new Order()
             {
-                OrderId = Convert.ToInt16(reader["OrderId"]),
-                ClientId = Convert.ToInt16(reader["ClientId"]),
+                OrderId = Convert.ToInt32(reader["OrderId"]),
+                ClientId = Convert.ToInt32(reader["ClientId"]),
                 OrderDate = Convert.ToString(reader["OrderDate"]),
                 OrderPayet = Convert.ToBoolean(reader["OrderPayet"])
             };
